fix: update `ownerPid` from `uint32` to `uint64` the same width as `size_t` on amd64/arm64 platform

+++ pkg/monitor/nvidia/v1/spec.go
@@ -63,7 +63,7 @@ type sharedRegionT struct {
 	majorVersion    int32
 	minorVersion    int32
 	smInitFlag      int32
-	ownerPid        uint32
+	ownerPid        uint64
 	sem             semT
 	num             uint64
 	uuids           [16]uuid
@@ -220,7 +220,7 @@ type sharedRegionTWithSemPostinit struct {
 	majorVersion      int32
 	minorVersion      int32
 	smInitFlag        int32
-	ownerPid          uint32
+	ownerPid          uint64
 	sem               semT
 	num               uint64
 	uuids             [16]uuid
